# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignmentLease": {
        "description": "Exclusive write lease for one delegate session.",
        "properties": {
          "expired_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expired At"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "heartbeat_at": {
            "format": "date-time",
            "title": "Heartbeat At",
            "type": "string"
          },
          "id": {
            "default": "",
            "minLength": 0,
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "minLength": 1,
            "title": "Owner Id",
            "type": "string"
          },
          "recovery_policy": {
            "$ref": "#/components/schemas/LeaseRecoveryPolicy",
            "default": "abandon"
          },
          "released_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "renewal_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Renewal Count",
            "type": "integer"
          },
          "session_id": {
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "supersedes_owner_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Owner Id"
          }
        },
        "required": [
          "session_id",
          "owner_id",
          "heartbeat_at",
          "expires_at"
        ],
        "title": "AssignmentLease",
        "type": "object"
      },
...
      "Checkpoint": {
        "description": "Durable delegate recovery evidence captured during execution.",
        "properties": {
          "artifact_refs": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Refs",
            "type": "array"
          },
          "changed_files": {
            "items": {
              "type": "string"
            },
            "title": "Changed Files",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_plan": {
            "minLength": 1,
            "title": "Current Plan",
            "type": "string"
          },
          "failures_and_learnings": {
            "items": {
              "type": "string"
            },
            "title": "Failures And Learnings",
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "resume_prompt": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Prompt"
          },
          "session_id": {
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "test_commands": {
            "items": {
              "type": "string"
            },
            "title": "Test Commands",
            "type": "array"
          }
        },
        "required": [
          "id",
          "session_id",
          "created_at",
          "current_plan"
        ],
        "title": "Checkpoint",
        "type": "object"
      },
...
      "DelegateSessionSnapshot": {
        "description": "Durable readback for one delegate session and its evidence.",
        "properties": {
          "active_lease": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssignmentLease"
              },
              {
                "type": "null"
              }
            ]
          },
          "handoff_artifacts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HandoffArtifact"
            },
            "title": "Handoff Artifacts",
            "type": "array"
          },
          "latest_checkpoint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Checkpoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/DelegateSession"
          }
        },
        "required": [
          "session"
        ],
        "title": "DelegateSessionSnapshot",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandoffArtifact": {
        "description": "Typed evidence or work product handed from one delegate to another.",
        "properties": {
          "artifact_ref": {
            "minLength": 1,
            "title": "Artifact Ref",
            "type": "string"
          },
          "artifact_type": {
            "minLength": 1,
            "title": "Artifact Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "session_id": {
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "artifact_type",
          "artifact_ref",
          "summary"
        ],
        "title": "HandoffArtifact",
        "type": "object"
      },
...
      "LeaseRecoveryPolicy": {
        "description": "Recovery behavior once a session lease expires.",
        "enum": [
          "abandon",
          "recoverable",
          "takeover_allowed"
        ],
        "title": "LeaseRecoveryPolicy",
        "type": "string"
      },
...
      "WorkersStatusResponse": {
        "properties": {
          "queue_depth": {
            "title": "Queue Depth",
            "type": "integer"
          },
          "worker_count": {
            "title": "Worker Count",
            "type": "integer"
          }
        },
        "required": [
          "worker_count",
          "queue_depth"
        ],
        "title": "WorkersStatusResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Maxwell-Daemon",
      "url": "https://github.com/D-sorganization/Maxwell-Daemon"
    },
    "description": "Maxwell-Daemon exposes a stable HTTP + WebSocket API for orchestrating agent tasks, observing pipeline state, and controlling the daemon lifecycle.\n\n**Contract version:** `2.0.0` (advertised at `GET /api/version`).  The contract is **append-only**: new fields and endpoints may appear, but existing request and response shapes are stable within a major version.\n\n## Authentication\n\nMost endpoints require a bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <your_token>\n```\n\nWhen JWT auth is configured, role-based access control (viewer / operator / admin) is enforced per endpoint.\n\n## Interactive docs\n\n* Swagger UI: [`/docs`](/docs)\n* ReDoc: [`/redoc`](/redoc)\n* OpenAPI JSON: [`/openapi.json`](/openapi.json)\n",
    "license": {
      "name": "MIT",
      "url": "https://github.com/D-sorganization/Maxwell-Daemon/blob/main/LICENSE"
    },
    "summary": "Autonomous AI control plane orchestrating agent tasks.",
    "title": "Maxwell-Daemon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/delegate-sessions": {
      "get": {
        "description": "List durable delegate sessions and their latest recovery evidence.",
        "operationId": "list_delegate_sessions_api_v1_delegate_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "delegate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delegate Id"
            }
          },
          {
            "in": "query",
            "name": "work_item_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Item Id"
            }
          },
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DelegateSessionSnapshot"
                  },
                  "title": "Response List Delegate Sessions Api V1 Delegate Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Delegate Sessions"
      }
    },
    "/api/v1/delegate-sessions/{session_id}": {
      "get": {
        "description": "Return one durable delegate session snapshot.",
        "operationId": "get_delegate_session_api_v1_delegate_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegateSessionSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Delegate Session"
      }
    },
...
    "/api/v1/fleet": {
      "get": {
        "description": "Return fleet manifest data merged with live task counts per repo.",
        "operationId": "fleet_overview_api_v1_fleet_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fleet Overview Api V1 Fleet Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fleet Overview"
      }
    },
    "/api/v1/fleet/capabilities": {
      "get": {
        "description": "Return a redacted capability registry snapshot for dispatch decisions.",
        "operationId": "fleet_capabilities_api_v1_fleet_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "repo",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Repo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tool",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Tool",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "required_capability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Required Capability"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fleet Capabilities Api V1 Fleet Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fleet Capabilities"
      }
    },
    "/api/v1/fleet/nodes": {
      "get": {
        "description": "Return a redacted capability registry snapshot for dispatch decisions.",
        "operationId": "fleet_capabilities_api_v1_fleet_nodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "repo",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Repo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tool",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Tool",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "required_capability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Required Capability"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fleet Capabilities Api V1 Fleet Nodes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fleet Capabilities"
      }
    },
    "/api/v1/heartbeat": {
      "post": {
        "description": "Workers POST here every heartbeat_seconds to stay registered as alive.",
        "operationId": "worker_heartbeat_api_v1_heartbeat_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Worker Heartbeat Api V1 Heartbeat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Heartbeat"
      }
    },
...
    "/api/v1/push/subscribe": {
      "post": {
        "description": "Register a Web Push subscription.\n\nCurrently a stub. Real implementation requires storing the subscription\nand generating a VAPID keypair.",
        "operationId": "push_subscribe_api_v1_push_subscribe_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Push Subscribe Api V1 Push Subscribe Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Subscribe"
      }
    },
...
    "/api/v1/workers": {
      "get": {
        "description": "Return current worker count and queue depth.",
        "operationId": "workers_status_api_v1_workers_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkersStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workers Status"
      },
      "put": {
        "description": "Rescale the worker pool to *count* workers.",
        "operationId": "set_workers_api_v1_workers_put",
        "parameters": [
          {
            "in": "query",
            "name": "count",
            "required": true,
            "schema": {
              "title": "Count",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Workers Api V1 Workers Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Workers"
      }
    },
...
  },
...
}